Move timestamp expressions to timestampExpressions.scala
